build: fix test for `signal` option so it works across Node.js versions
We test the `signal` option by passing in a value that isn't a
signal and expecting `fetch` to throw an error.

An error is thrown, but our current assertion doesn't work in
Node.js v18 because the error message has changed. This changes
the assertion to work with the behaviour of Node v14, v16 and v18.